pyproject.toml
[tool.poetry]
name = "ochra"
version = "0.1.0"
description = "Declarative, composable, & stateless drawing"
authors = ["Tongfei Chen <tongfei@pm.me>"]
readme = "README.md"
repository = "https://github.com/ctongfei/ochra"
packages = [{ include = "ochra", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10"
scipy = "^1.11.2"
numpy = "^1.25.2"
pycairo = "^1.26.0"
cairosvg = "^2.7.1"

[tool.poetry.group.test.dependencies]
vega-datasets = "^0.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[tool.ruff.lint]
ignore = ["E731", "E741"]
